+++ src/deploy/table_event_format_mapping.sql
@@ -14,8 +14,11 @@ COMMENT ON TABLE vibetype.event_format_mapping IS 'Mapping events to formats (M:
 COMMENT ON COLUMN vibetype.event_format_mapping.event_id IS 'An event id.';
 COMMENT ON COLUMN vibetype.event_format_mapping.format_id IS 'A format id.';
 
+\set role_service_reccoom_username `cat /run/secrets/postgres_role_service_reccoom_username`
+
 GRANT SELECT ON TABLE vibetype.event_format_mapping TO vibetype_anonymous;
 GRANT SELECT, INSERT, DELETE ON TABLE vibetype.event_format_mapping TO vibetype_account;
+GRANT SELECT ON TABLE vibetype.event_format_mapping TO :role_service_reccoom_username;
 
 ALTER TABLE vibetype.event_format_mapping ENABLE ROW LEVEL SECURITY;