+++ src/verify/table_account_block.sql
@@ -10,6 +10,8 @@ ROLLBACK;
 
 BEGIN;
 
+\set role_service_reccoom_username `cat /run/secrets/postgres_role_service_reccoom_username`
+SET local role.reccoom_username TO :'role_service_reccoom_username';
 \set role_service_vibetype_username `cat /run/secrets/postgres_role_service_vibetype_username`
 SET local role.vibetype_username TO :'role_service_vibetype_username';
 
@@ -23,6 +25,10 @@ BEGIN
   ASSERT NOT (SELECT pg_catalog.has_table_privilege('vibetype_anonymous', 'vibetype.account_block', 'INSERT'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege('vibetype_anonymous', 'vibetype.account_block', 'UPDATE'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege('vibetype_anonymous', 'vibetype.account_block', 'DELETE'));
+  ASSERT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.account_block', 'SELECT'));
+  ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.account_block', 'INSERT'));
+  ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.account_block', 'UPDATE'));
+  ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.account_block', 'DELETE'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.vibetype_username'), 'vibetype.account_block', 'SELECT'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.vibetype_username'), 'vibetype.account_block', 'INSERT'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.vibetype_username'), 'vibetype.account_block', 'UPDATE'));