+++ src/verify/table_preference_event_format.sql
@@ -13,6 +13,8 @@ ROLLBACK;
 
 BEGIN;
 
+\set role_service_reccoom_username `cat /run/secrets/postgres_role_service_reccoom_username`
+SET local role.reccoom_username TO :'role_service_reccoom_username';
 \set role_service_vibetype_username `cat /run/secrets/postgres_role_service_vibetype_username`
 SET local role.vibetype_username TO :'role_service_vibetype_username';
 
@@ -26,6 +28,10 @@ BEGIN
   ASSERT NOT (SELECT pg_catalog.has_table_privilege('vibetype_anonymous', 'vibetype.preference_event_format', 'INSERT'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege('vibetype_anonymous', 'vibetype.preference_event_format', 'UPDATE'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege('vibetype_anonymous', 'vibetype.preference_event_format', 'DELETE'));
+  ASSERT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.preference_event_format', 'SELECT'));
+  ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.preference_event_format', 'INSERT'));
+  ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.preference_event_format', 'UPDATE'));
+  ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.reccoom_username'), 'vibetype.preference_event_format', 'DELETE'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.vibetype_username'), 'vibetype.preference_event_format', 'SELECT'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.vibetype_username'), 'vibetype.preference_event_format', 'INSERT'));
   ASSERT NOT (SELECT pg_catalog.has_table_privilege(current_setting('role.vibetype_username'), 'vibetype.preference_event_format', 'UPDATE'));